Black formatting

+++ research/strategies/pairs_trading/pairs.py
@@ -1,4 +1,4 @@
-import pandas as pd 
+import pandas as pd
 import numpy as np
 import matplotlib.pyplot as plt
 from itertools import combinations
@@ -13,50 +13,66 @@ YEARLY_PAIRS_PATH = DATA_DIR + "/yearly_pairs.csv"
 
 crsp = CRSP()
 
-df = crsp.df[['permno', 'date', 'ticker', 'prc', 'ret']].copy()
+df = crsp.df[["permno", "date", "ticker", "prc", "ret"]].copy()
 
 df = df.query("prc > 5").reset_index(drop=True)
 
 # Add extra date variables
-df['mdt'] = df['date'].dt.strftime("%Y-%m")
-df['year'] = df['date'].dt.strftime("%Y")
-df['month'] = df['date'].dt.strftime("%m")
+df["mdt"] = df["date"].dt.strftime("%Y-%m")
+df["year"] = df["date"].dt.strftime("%Y")
+df["month"] = df["date"].dt.strftime("%m")
 
 # Calculate Different Return Variables
 
-holding_period = 6 # Alternative is 1
-holding_period_var = f'ret_{holding_period}'
+holding_period = 6  # Alternative is 1
+holding_period_var = f"ret_{holding_period}"
 
 # Log Returns
-df['logret'] = np.log1p(df['ret'])
-df['cumret'] = df.groupby(['permno','year'])['logret'].cumsum().reset_index(drop=True)
-df['cumret_lag'] = df.groupby('permno')['cumret'].shift(1)
+df["logret"] = np.log1p(df["ret"])
+df["cumret"] = df.groupby(["permno", "year"])["logret"].cumsum().reset_index(drop=True)
+df["cumret_lag"] = df.groupby("permno")["cumret"].shift(1)
 
 # Holding period returns
-df[holding_period_var] = df.groupby('permno')['logret'].rolling(holding_period,holding_period).sum().reset_index(drop=True)
-df[holding_period_var] = df.groupby('permno')[holding_period_var].shift(-(holding_period-1))
+df[holding_period_var] = (
+    df.groupby("permno")["logret"]
+    .rolling(holding_period, holding_period)
+    .sum()
+    .reset_index(drop=True)
+)
+df[holding_period_var] = df.groupby("permno")[holding_period_var].shift(
+    -(holding_period - 1)
+)
 
-pivot = df.groupby(['permno','date'])['cumret_lag'].mean().reset_index().pivot(index='date', columns='permno', values='cumret_lag')
-pivot['year'] = pivot.index.year
+pivot = (
+    df.groupby(["permno", "date"])["cumret_lag"]
+    .mean()
+    .reset_index()
+    .pivot(index="date", columns="permno", values="cumret_lag")
+)
+pivot["year"] = pivot.index.year
 pivot
 
-years = pivot['year'].unique()  # [2017, 2018]
+years = pivot["year"].unique()  # [2017, 2018]
 
 mse_frame = pd.DataFrame(index=years, columns=pivot.columns[:-1], data=float("inf"))
 pairs_frame = pd.DataFrame(index=years, columns=pivot.columns[:-1], data=None)
 
 # Iterated through each year of data
-for year in tqdm(years, desc='Processing years'):
-    slice = pivot[pivot['year'] == year].dropna(axis=1, how='all').drop(columns=['year'])
-    
-    returns = slice.dropna(axis=1).T.dropna().T # This gets rid of stocks without 12 months of returns
+for year in tqdm(years, desc="Processing years"):
+    slice = (
+        pivot[pivot["year"] == year].dropna(axis=1, how="all").drop(columns=["year"])
+    )
+
+    returns = (
+        slice.dropna(axis=1).T.dropna().T
+    )  # This gets rid of stocks without 12 months of returns
     returns_array = returns.values
-    
+
     # Loop through all combination of pairs
     for i, j in combinations(range(returns_array.shape[1]), 2):
         stock_i = returns.columns[i]
         stock_j = returns.columns[j]
-        
+
         stock_i_returns = returns_array[:, i]
         stock_j_returns = returns_array[:, j]
 
@@ -74,16 +90,28 @@ pairs_frame = pd.read_csv(PAIRS_PATH, index_col=0)
 mse_frame = pd.read_csv(MSE_PATH, index_col=0)
 
 # Reformat mse and pairs dataframes
-result_pairs_frame = pairs_frame.unstack().reset_index().rename(columns={'level_0': 'permno','level_1': 'year', 0: 'pair'})
-result_mse_frame = mse_frame.unstack().reset_index().rename(columns={'level_0': 'permno','level_1':'year', 0: 'mse'})
+result_pairs_frame = (
+    pairs_frame.unstack()
+    .reset_index()
+    .rename(columns={"level_0": "permno", "level_1": "year", 0: "pair"})
+)
+result_mse_frame = (
+    mse_frame.unstack()
+    .reset_index()
+    .rename(columns={"level_0": "permno", "level_1": "year", 0: "mse"})
+)
 
 # Merge
-merged = result_mse_frame.merge(result_pairs_frame, on=['permno','year'], how='left')
+merged = result_mse_frame.merge(result_pairs_frame, on=["permno", "year"], how="left")
 
 # Clean up final dataframe
-merged = merged.sort_values(by=['year','mse'], ascending=True).dropna().reset_index(drop=True)
+merged = (
+    merged.sort_values(by=["year", "mse"], ascending=True)
+    .dropna()
+    .reset_index(drop=True)
+)
 
-merged['permno'] = merged['permno'].astype(int)
-merged['pair'] = merged['pair'].astype(int)
+merged["permno"] = merged["permno"].astype(int)
+merged["pair"] = merged["pair"].astype(int)
 
 merged.to_csv(YEARLY_PAIRS_PATH, index=False)
